backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
redis==5.0.1
pydantic>=2.9.2,<3
pydantic-settings>=2.3.0,<3
pytest==7.4.3
